backend/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
protobuf==3.19.6
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
